Successfully bubbling out errors

+++ contracts/DelegatableCore.sol
@@ -55,19 +55,22 @@ abstract contract DelegatableCore is EIP712Decoder {
         }
     }
 
-    function extractRevertReason (bytes memory revertData)
-        internal pure returns (string memory reason) {
-        uint l = revertData.length;
+    function extractRevertReason(bytes memory revertData)
+        internal
+        pure
+        returns (string memory reason)
+    {
+        uint256 l = revertData.length;
         if (l < 68) return "";
-        uint t;
+        uint256 t;
         assembly {
-            revertData := add (revertData, 4)
-            t := mload (revertData) // Save the content of the length slot
-            mstore (revertData, sub (l, 4)) // Set proper length
+            revertData := add(revertData, 4)
+            t := mload(revertData) // Save the content of the length slot
+            mstore(revertData, sub(l, 4)) // Set proper length
         }
-        reason = abi.decode (revertData, (string));
+        reason = abi.decode(revertData, (string));
         assembly {
-            mstore (revertData, t) // Restore the content of the length slot
+            mstore(revertData, t) // Restore the content of the length slot
         }
     }
